+++ types/txresponse.go
@@ -1,10 +1,10 @@
 package types
 
 import (
-	"github.com/cosmos/relayer/v2/relayer/provider"
+	"github.com/babylonlabs-io/babylon/client/babylonclient"
 )
 
 type TxResponse struct {
 	TxHash string
-	Events []provider.RelayerEvent
+	Events []babylonclient.RelayerEvent
 }